Mnor bugfix for 64bit cpu
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@23331 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSProcessInfo.m
@@ -934,7 +934,7 @@ int main(int argc, char *argv[], char *env[])
 - (NSString *) globallyUniqueString
 {
   static unsigned long	counter = 0;
-  int			count;
+  unsigned long		count;
   static NSString	*host = nil;
   static int		pid;
   static unsigned long	start;
